修正一堆嚴重的錯誤。 如：已經有人開始選課，就不能再刪除課程。 修正每次更新一次 job 裡面的課程就全部刪除重建。

+++ src/tw/jiangsir/ZeroJiaowu/Servlets/Do/DoCourseServlet.java
@@ -6,7 +6,10 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.*;
 
 import tw.jiangsir.Utils.Annotations.RoleSetting;
+import tw.jiangsir.Utils.Exceptions.DataException;
 import tw.jiangsir.ZeroJiaowu.DAOs.CourseDAO;
+import tw.jiangsir.ZeroJiaowu.DAOs.ElectiveDAO;
+import tw.jiangsir.ZeroJiaowu.Objects.Course;
 import tw.jiangsir.ZeroJiaowu.Objects.User.ROLE;
 
 @SuppressWarnings("serial")
@@ -20,6 +23,11 @@ public class DoCourseServlet extends HttpServlet {
 		String action = request.getParameter("action");
 		if ("deleteCourse".equals(action)) {
 			int courseid = Integer.parseInt(request.getParameter("courseid"));
+			Course course = new CourseDAO().getCourseById(courseid);
+			if (new ElectiveDAO().getElectivesByJobid(course.getJobid()).size() > 0) {
+				throw new DataException(
+						"本選課作業(" + course.getJob().getTitle() + ")已經有人開始選課了，不能刪除任何一個課程（" + course.getName() + "）。");
+			}
 			new CourseDAO().delete(courseid);
 		}
 		response.getWriter().print("Done");
